Fixed the failing StandingsParser test. Updated the JUnit version to 4.8.2 and updated the tests to reflect that assertEquals(double expected, double actual) now fails in favour of using assertEquals(double expected, double actual, delta)

+++ trunk/src/test/java/com/beimin/eveapi/character/mailmessages/MailMessagesParserTest.java
@@ -39,7 +39,7 @@ public class MailMessagesParserTest {
 				calendar.set(Calendar.MILLISECOND, 0);
 				assertEquals(calendar.getTime(), mail.getSentDate());
 				assertEquals("FW: hulkageddon 2", mail.getTitle());
-				assertEquals(1449814438L, mail.getToCorpOrAllianceID());
+				assertEquals(1449814438L, mail.getToCorpOrAllianceID().longValue());
 				assertNull(mail.getToCharacterIDs());
 				assertNull(mail.getToListIDs());
 				assertEquals(false, mail.isRead());
